Move share text to server response
